Adjust logical ordering of ROI extraction. Duplicate one line of code for readability.

+++ clpipe/roi_extractor.py
@@ -280,9 +280,10 @@ def _fmri_roi_extract_subject(
         ),
         exist_ok=True,
     )
-    shutil.copy2(
-        atlas_labelpath, config.config["ROIExtractionOptions"]["OutputDirectory"]
-    )
+    if not Path(atlas_labelpath).exists():
+        shutil.copy2(
+            atlas_labelpath, config.config["ROIExtractionOptions"]["OutputDirectory"]
+        )
 
     for file in subject_files:
         fmri_roi_extract_image(
@@ -326,12 +327,9 @@ def fmri_roi_extract_image(
         logger.info("File Exists! Skipping. Use -overwrite to reprocess.")
         return
 
-    found_mask = False
-    fallback = False
     try:
         # First, try to find this image's mask.
         mask_file = _mask_finder(file, config, logger)
-        found_mask = True
     except MaskFileNotFoundError:
         if config.config["ROIExtractionOptions"]["RequireMask"]:
             # If a mask is required, return here due to missing mask.
@@ -341,6 +339,9 @@ def fmri_roi_extract_image(
             # Otherwise, continue on with a warning.
             logger.warning(
                 "Unable to find a mask for this image. Extracting ROIs without using brain mask."
+            )
+            ROI_ts = _fmri_roi_extract_image(
+                file, atlas_path, atlas_type, radius, overlap_ok, logger
             )
     else:
         try:
@@ -358,17 +359,11 @@ def fmri_roi_extract_image(
         except ValueError as ve:
             # Trigger fallback flag if any ROIs are outside of the mask region.
             logger.warning(ve.__str__() + ". Extracting ROIs without using brain mask.")
-            fallback = True
+            logger.info("Starting non-masked ROI extraction...")
+            ROI_ts = _fmri_roi_extract_image(
+                file, atlas_path, atlas_type, radius, overlap_ok, logger
+            )
 
-    # Run non-masked extraction if no mask found or fallback needed
-    if not found_mask or fallback:
-        logger.info("Starting non-masked ROI extraction...")
-        ROI_ts = _fmri_roi_extract_image(
-            file, atlas_path, atlas_type, radius, overlap_ok, logger
-        )
-
-    
-    if fallback:
         temp_mask = concat_imgs([mask_file, mask_file])
         mask_ROIs = _fmri_roi_extract_image(
             temp_mask, atlas_path, atlas_type, radius, overlap_ok, logger
